Changed `BPM` type from `short` to `uint` and updated validation logic to enforce non-negative values. Added test for negative BPM input.

+++ Listen2MeRefined.Core/Models/AudioModel.cs
@@ -14,7 +14,7 @@ public sealed class AudioModel : ModelBase
 
     public string? Genre { get; set; }
 
-    public short BPM { get; set; }
+    public uint BPM { get; set; }
 
     public short Bitrate { get; set; }
